refactor: create fare-contract-validity-units.ts
