+++ src/modules/fare-contracts/__tests__/fare-contract-validity-units.test.ts
@@ -1,4 +1,4 @@
-import {fareContractValidityUnits} from '../utils';
+import {fareContractValidityUnits} from '../fare-contract-validity-units';
 import {LoadingParams} from '@atb/loading-screen/types';
 import React from 'react';
